% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_ai_mod}
\alias{calculate_ai_mod}
\title{Calculate modified aromaticity index}
\usage{
calculate_ai_mod(data)
}
\arguments{
\item{data}{Formula data frame.}
}
\value{
Numeric AI_mod vector.
}
\description{
Uses the commonly applied DOM formula expression. Values are set to zero
when the numerator is non-positive or denominator is non-positive.
Interpretation for P- and S-containing formulae requires caution.
}
\references{
Koch BP, Dittmar T (2006). \doi{10.1002/rcm.2386}
}
