% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_formula_metrics}
\alias{calculate_formula_metrics}
\title{Calculate all formula-level DOM indices}
\usage{
calculate_formula_metrics(data)
}
\arguments{
\item{data}{Formula data frame containing CHONPS atom counts.}
}
\value{
Input data with elemental ratios, DBE, DBE/C, DBE/O, AI_mod,
and NOSC.
}
\description{
Calculate all formula-level DOM indices
}
\examples{
calculate_formula_metrics(data.frame(C = 10, H = 12, O = 5, N = 0, P = 0, S = 0))
}
