% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{calculate_shared_formulas}
\alias{calculate_shared_formulas}
\title{Calculate shared molecular formulae}
\usage{
calculate_shared_formulas(data, min_samples = 1L)
}
\arguments{
\item{data}{Assigned formula table.}

\item{min_samples}{Minimum number of samples in which a formula must occur.}
}
\value{
Formula occurrence table with a semicolon-separated sample set.
}
\description{
Calculate shared molecular formulae
}
