% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{default_vk_regions}
\alias{default_vk_regions}
\title{Default mutually exclusive van Krevelen regions}
\usage{
default_vk_regions()
}
\value{
Data frame of ordered region boundaries.
}
\description{
Default mutually exclusive van Krevelen regions
}
