% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{filter_replicate_peaks}
\alias{filter_replicate_peaks}
\title{Filter peaks by replicate occurrence}
\usage{
filter_replicate_peaks(data, min_fraction = 2/3, ppm = 1)
}
\arguments{
\item{data}{Standardized peak table with \code{replicate_id}.}

\item{min_fraction}{Minimum fraction of replicates within each sample group.}

\item{ppm}{Matching tolerance.}
}
\value{
Peaks retained in the requested replicate fraction.
}
\description{
Filter peaks by replicate occurrence
}
