% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_formula_metrics}
\alias{plot_formula_metrics}
\title{Plot formula-metric distributions}
\usage{
plot_formula_metrics(
  data,
  metric = "DBE",
  group = NULL,
  bins = 35,
  facet = NULL,
  theme = NULL,
  return_data = FALSE
)
}
\arguments{
\item{data}{Assigned formula table.}

\item{metric}{One of H_C, O_C, DBE, AI_mod, or NOSC.}

\item{group}{Optional grouping column.}

\item{bins}{Histogram bins.}

\item{facet}{Optional faceting column.}

\item{theme}{ggplot2 theme.}

\item{return_data}{Return plot and data.}
}
\value{
A ggplot object or list.
}
\description{
Plot formula-metric distributions
}
