% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mass_error}
\alias{plot_mass_error}
\title{Plot mass errors}
\usage{
plot_mass_error(
  data,
  distribution = FALSE,
  group = NULL,
  alpha = 0.6,
  point_size = 1.2,
  xlim = NULL,
  ylim = NULL,
  facet = NULL,
  theme = NULL,
  return_data = FALSE
)
}
\arguments{
\item{data}{Assigned formula table.}

\item{distribution}{Draw a histogram instead of m/z versus mass error.}

\item{group}{Optional grouping column.}

\item{alpha}{Point transparency.}

\item{point_size}{Point size.}

\item{xlim, ylim}{Optional coordinate limits.}

\item{facet}{Optional faceting column.}

\item{theme}{ggplot2 theme object.}

\item{return_data}{Return a list with plot and plot data.}
}
\value{
A ggplot object or list.
}
\description{
Plot mass errors
}
