% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_molecular_richness}
\alias{plot_molecular_richness}
\title{Plot molecular richness}
\usage{
plot_molecular_richness(data, theme = NULL, return_data = FALSE)
}
\arguments{
\item{data}{Assigned formula table.}

\item{theme}{ggplot2 theme.}

\item{return_data}{Return plot and data.}
}
\value{
A ggplot object or list.
}
\description{
Plot molecular richness
}
