% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sample_composition}
\alias{plot_sample_composition}
\title{Plot sample composition}
\usage{
plot_sample_composition(
  data,
  weight = c("count", "intensity"),
  position = "stack",
  theme = NULL,
  return_data = FALSE
)
}
\arguments{
\item{data}{Assigned formula table.}

\item{weight}{\code{count} or \code{intensity}.}

\item{position}{Bar position.}

\item{theme}{ggplot2 theme.}

\item{return_data}{Return plot and summarized data.}
}
\value{
A ggplot object or list.
}
\description{
Plot sample composition
}
