% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_dom_config}
\alias{read_dom_config}
\title{Read a YAML workflow configuration}
\usage{
read_dom_config(path)
}
\arguments{
\item{path}{YAML file path.}
}
\value{
A validated \code{dom_config}.
}
\description{
Read a YAML workflow configuration
}
